+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/HighPassCounterMapperRegistry.java
@@ -14,6 +14,8 @@ import org.springframework.stereotype.Component;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 /**
  * 高通计数映射器
@@ -79,10 +81,49 @@ public class HighPassCounterMapperRegistry extends AbstractMapperRegistry {
 
         @Override
         protected Object doAggregate(MapParam mapParam, List<Item> items, Date startDate, Date endDate) {
-            Sequence s = HighPassMapperRegistry.HighPassMapper.highPass(
-                    mapParam, new Sequence(new LongIdKey(0L), items, startDate, endDate)
-            );
+            Sequence s = highPass(mapParam, new Sequence(new LongIdKey(0L), items, startDate, endDate));
             return s.getItems().size();
+        }
+
+        @SuppressWarnings("DuplicatedCode")
+        private Sequence highPass(MapParam mapParam, Sequence sequence) {
+            // 解析配置。
+            Config config = JSON.parseObject(mapParam.getParam(), Config.class);
+
+            // 展开参数。
+            double threshold = config.getThreshold();
+            boolean canEqual = config.isCanEqual();
+            boolean invert = config.isInvert();
+
+            // 定义数据条目列表，并进行过滤。
+            List<Item> items = doFilter(sequence, threshold, canEqual, invert);
+
+            // 返回结果。
+            return new Sequence(sequence.getPointKey(), items, sequence.getStartDate(), sequence.getEndDate());
+        }
+
+        // 为了保证代码的可读性，此处代码不做简化。
+        @SuppressWarnings({"ConstantValue", "DuplicatedCode"})
+        private static List<Item> doFilter(Sequence sequence, double threshold, boolean canEqual, boolean invert) {
+            List<Item> items;
+            if (invert && canEqual) {
+                items = sequence.getItems().stream().filter(
+                        item -> (Objects.isNull(item.getValue()) ? 0.00 : (double) item.getValue()) <= threshold
+                ).collect(Collectors.toList());
+            } else if (invert && !canEqual) {
+                items = sequence.getItems().stream().filter(
+                        item -> (Objects.isNull(item.getValue()) ? 0.00 : (double) item.getValue()) < threshold
+                ).collect(Collectors.toList());
+            } else if (!invert && canEqual) {
+                items = sequence.getItems().stream().filter(
+                        item -> (Objects.isNull(item.getValue()) ? 0.00 : (double) item.getValue()) >= threshold
+                ).collect(Collectors.toList());
+            } else {
+                items = sequence.getItems().stream().filter(
+                        item -> (Objects.isNull(item.getValue()) ? 0.00 : (double) item.getValue()) > threshold
+                ).collect(Collectors.toList());
+            }
+            return items;
         }
 
         @Override